backend/requirements.txt
fastapi==0.123.4
uvicorn==0.38.0
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyPDF2==3.0.1
google-genai==1.52.0